refactor(cli): extract manifest provider domain
